requirements.txt
boto3==1.37.13
python-dotenv==1.0.1
Requests==2.32.3
